chore: refactor cli migrate command into focused modules
